Attempt to fix Windows build.
Review URL: http://codereview.chromium.org/14920

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@7345 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/net/dns_master.h
@@ -151,7 +151,7 @@ class DnsMaster {
   bool PreLockedCreateNewSlaveIfNeeded();  // Lazy slave processes creation.
 
   // Number of slave processes started early (to help with startup prefetch).
-  static const int kSlaveCountMin = 4;
+  static const size_t kSlaveCountMin = 4;
 
   // Synchronize access to results_, referrers_, and slave control data.
   Lock lock_;
